Add files via upload
As Decision tree is sensitive to noise and overfits when the data is sparse, the training and testing was not promising. However, the RMSE was much better than Linear regression on data of the year 2022. I also tried to prevent overfitting by the following techniques:
1- Rescaling the features values (Mean= 0, Variance=1)
2- Dimensionality reduction using PCA. I tried to find the optimum principal components(there is a graph related to it).
3- Reducing the Max_Dept of the decision Tree.

Now, I have combined data from 2019 to 2022 which have mutual features in order to find relationships between features same as before.(for example to predict color based on, Ph, Conductivity, Turbidity, Chlorine, and alkalinity)
I cleaned and preprocessed the data.(It has some commas or other string values).